% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{collapse_across_clusters}
\alias{collapse_across_clusters}
\title{Collapse Across Clusters}
\usage{
collapse_across_clusters(x, collapse_fun = NULL)
}
\arguments{
\item{x}{Values to collapse}

\item{collapse_fun}{Function to use for collapsing}
}
\value{
Collapsed values
}
\description{
Helper function to collapse values across clusters
}
\keyword{internal}
