% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_heatmap_colors}
\alias{create_heatmap_colors}
\title{Create Heatmap Colors}
\usage{
create_heatmap_colors(n = 100)
}
\arguments{
\item{n}{Number of colors to generate}
}
\value{
Character vector of colors
}
\description{
Generates a color palette for heatmap visualization
}
\keyword{internal}
