% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rinvGauss}
\alias{.rinvGauss}
\title{Draw from Inverse Gaussian Distribution}
\usage{
.rinvGauss(n, mu, lambda)
}
\arguments{
\item{n}{Number of observations}

\item{mu}{Mean parameter}

\item{lambda}{Shape parameter}
}
\value{
Vector of random numbers
}
\description{
Generate random numbers from inverse Gaussian distribution
}
\keyword{internal}
