% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_mrd}
\alias{format_mrd}
\title{Format MRD Value}
\usage{
format_mrd(x)
}
\arguments{
\item{x}{Numeric MRD value}
}
\value{
Character string with formatted MRD value
}
\description{
Formats MRD values for display
}
\keyword{internal}
