% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_time}
\alias{format_time}
\title{Format Time}
\usage{
format_time(x)
}
\arguments{
\item{x}{Numeric time value}
}
\value{
Character string with formatted time
}
\description{
Formats time values for display
}
\keyword{internal}
