% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_color_map}
\alias{generate_color_map}
\title{Generate Color Map}
\usage{
generate_color_map(values)
}
\arguments{
\item{values}{Vector of unique values to map to colors}
}
\value{
Named vector of colors
}
\description{
Generates a color mapping for categorical variables
}
\keyword{internal}
