% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.scalar}
\alias{is.scalar}
\title{Check Scalar}
\usage{
is.scalar(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if x is a numeric scalar
}
\description{
Check if x is a numeric scalar
}
\keyword{internal}
