% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_binary}
\alias{is_binary}
\title{Check Binary Variable}
\usage{
is_binary(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
Logical indicating whether x is binary
}
\description{
Checks if a variable is binary (0/1 or TRUE/FALSE)
}
\keyword{internal}
