% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{jm_hazard}
\alias{jm_hazard}
\title{Calculate Hazard for Joint Model}
\usage{
jm_hazard(
  t,
  x,
  betas,
  basehaz = "weibull",
  M = 1,
  trajectory = "mrd_piecewise",
  family = list("lognormal"),
  grp_assoc = NULL
)
}
\arguments{
\item{t}{Time point}

\item{x}{Covariates}

\item{betas}{Model parameters}

\item{basehaz}{Baseline hazard type}

\item{M}{Number of markers}

\item{trajectory}{Trajectory type}

\item{family}{Distribution family}

\item{grp_assoc}{Group association structure}
}
\value{
Calculated hazard value
}
\description{
Calculates the hazard function for the joint model
}
\keyword{internal}
