% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{lognormal}
\alias{lognormal}
\title{Lognormal Family}
\usage{
lognormal()
}
\value{
A family object with the following components:
\item{family}{Character string "lognormal"}
\item{link}{Link function name "log"}
\item{linkinv}{Inverse link function (exponential)}
\item{variance}{Variance function}
\item{dev.resids}{Deviance residuals function}
\item{aic}{AIC function}
\item{initialize}{Initialization expression}
}
\description{
Creates a family object for lognormal distribution, used in the simulation of MRD data.
This is a custom family object that implements the lognormal distribution for use
in modeling longitudinal MRD measurements.
}
