% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{maybe_broadcast}
\alias{maybe_broadcast}
\title{Maybe Broadcast}
\usage{
maybe_broadcast(x, n)
}
\arguments{
\item{x}{A vector or scalar}

\item{n}{Number of replications to possibly make}
}
\value{
Broadcasted vector
}
\description{
Broadcast a vector or scalar to specified length
}
\keyword{internal}
