% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_mrdviz.R
\name{mrdviz_server}
\alias{mrdviz_server}
\title{MRDviz Server Functions}
\usage{
mrdviz_server(input, output, session, shared_storage)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{shared_storage}{Reactive values object for storing shared data}
}
\description{
Server-side logic for the MRDviz application
}
