% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlist}
\alias{nlist}
\title{Create Named List}
\usage{
nlist(...)
}
\arguments{
\item{...}{Objects to include in the list}
}
\value{
A named list
}
\description{
Create a named list using specified names or object names
}
\keyword{internal}
