% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{plot_all_trajectories}
\alias{plot_all_trajectories}
\title{Plot All Trajectories}
\usage{
plot_all_trajectories(sim_data_list, show_true = FALSE)
}
\arguments{
\item{sim_data_list}{List of simulated data for each group}

\item{show_true}{Logical; whether to show true trajectories}
}
\value{
A ggplot2 object
}
\description{
Creates a plot of MRD trajectories for all groups
}
