% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{plot_trajectories}
\alias{plot_trajectories}
\title{Plot Trajectories}
\usage{
plot_trajectories(sim_data, show_true = FALSE)
}
\arguments{
\item{sim_data}{Simulated data list}

\item{show_true}{Logical; whether to show true trajectories}
}
\value{
A ggplot2 object
}
\description{
Creates a plot of MRD trajectories
}
