% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mrdviz.R
\name{run_mrdviz}
\alias{run_mrdviz}
\title{Launch the MRDviz Application}
\usage{
run_mrdviz()
}
\description{
This function launches the MRDviz Shiny application, which provides interactive
visualization and simulation capabilities for Minimal Residual Disease (MRD) data
in clinical trials.
}
\examples{
if (interactive()) {
  run_mrdviz()
}
}
