% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simjm}
\alias{simjm}
\title{Simulate Joint Model for MRD Data}
\usage{
simjm(
  n = 200,
  M = 1,
  fixed_trajectory = "mrd_piecewise",
  random_trajectory = "linear",
  assoc = "etavalue",
  basehaz = c("weibull"),
  missing_data_rate = 0,
  betaLong_intercept = 10,
  betaLong_binary = -1,
  betaLong_continuous = 1,
  betaLong_linear = -0.25,
  betaLong_quadratic = 0.03,
  betaLong_cubic = -0.0015,
  betaLong_mrd_eot = 6,
  betaLong_mrd_relapse_prob = 0.3,
  betaLong_mrd_relapse_time_mean = 12,
  betaLong_mrd_relapse_time_sd = 2,
  betaLong_mrd_relapse_rate = 0.2,
  betaLong_baseline_mean = 4,
  betaLong_baseline_sd = 0.3,
  betaLong_plateau_mean = 1,
  betaLong_plateau_sd = 0.2,
  betaLong_plateau_binary = -0.1,
  betaLong_plateau_continuous = -0.05,
  betaLong_aux = 0.5,
  betaEvent_intercept = -7.5,
  betaEvent_binary = -0.5,
  betaEvent_continuous = 0.5,
  betaEvent_assoc_value = 0.5,
  betaEvent_assoc_slope = 0.5,
  betaEvent_assoc_auc = 0.5,
  betaEvent_aux = 1.2,
  b_sd = c(1.5, 0.07),
  b_rho = -0.2,
  prob_Z1 = 0.5,
  mean_Z2 = 0,
  sd_Z2 = 1,
  max_yobs = 10,
  max_fuptime = 20,
  balanced = FALSE,
  family = "lognormal",
  clust_control = list(),
  return_eta = FALSE,
  grp_assoc = NULL,
  seed = sample.int(.Machine$integer.max, 1),
  interval = c(1e-08, 200)
)
}
\arguments{
\item{n}{Number of subjects}

\item{M}{Number of markers}

\item{fixed_trajectory}{Type of fixed trajectory ("mrd_piecewise", "linear", "quadratic", "cubic")}

\item{random_trajectory}{Type of random trajectory}

\item{assoc}{Association structure}

\item{basehaz}{Baseline hazard type}

\item{missing_data_rate}{Proportion of missing data points (0-1)}

\item{betaLong_intercept}{Intercept for longitudinal component}

\item{betaLong_binary}{Binary covariate effect}

\item{betaLong_continuous}{Continuous covariate effect}

\item{betaLong_linear}{Linear time effect}

\item{betaLong_quadratic}{Quadratic time effect}

\item{betaLong_cubic}{Cubic time effect}

\item{betaLong_mrd_eot}{End of treatment time}

\item{betaLong_mrd_relapse_prob}{Relapse probability}

\item{betaLong_mrd_relapse_time_mean}{Mean relapse time}

\item{betaLong_mrd_relapse_time_sd}{SD of relapse time}

\item{betaLong_mrd_relapse_rate}{Relapse rate}

\item{betaLong_baseline_mean}{Baseline MRD mean}

\item{betaLong_baseline_sd}{Baseline MRD SD}

\item{betaLong_plateau_mean}{Plateau MRD mean}

\item{betaLong_plateau_sd}{Plateau MRD SD}

\item{betaLong_plateau_binary}{Binary covariate effect on plateau}

\item{betaLong_plateau_continuous}{Continuous covariate effect on plateau}

\item{betaLong_aux}{Auxiliary parameter for longitudinal component}

\item{betaEvent_intercept}{Event model intercept}

\item{betaEvent_binary}{Binary covariate effect on event}

\item{betaEvent_continuous}{Continuous covariate effect on event}

\item{betaEvent_assoc_value}{Association parameter for current value}

\item{betaEvent_assoc_slope}{Association parameter for slope}

\item{betaEvent_assoc_auc}{Association parameter for AUC}

\item{betaEvent_aux}{Auxiliary parameter for event model}

\item{b_sd}{Random effects SD}

\item{b_rho}{Random effects correlation}

\item{prob_Z1}{Probability for binary covariate}

\item{mean_Z2}{Mean for continuous covariate}

\item{sd_Z2}{SD for continuous covariate}

\item{max_yobs}{Maximum number of observations per subject}

\item{max_fuptime}{Maximum follow-up time}

\item{balanced}{Whether to use balanced design}

\item{family}{Distribution family}

\item{clust_control}{Clustering control parameters}

\item{return_eta}{Whether to return linear predictor}

\item{grp_assoc}{Group association structure}

\item{seed}{Random seed}

\item{interval}{Interval for hazard calculation}
}
\value{
A list containing simulated longitudinal and event data
}
\description{
Simulates longitudinal and survival data for MRD trajectories using a joint modeling approach
}
