% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_simulation.R
\name{simulation_server}
\alias{simulation_server}
\title{MRDviz Simulation Server Functions}
\usage{
simulation_server(input, output, session, shared_storage)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{shared_storage}{Reactive values object for storing shared data}
}
\description{
Server-side logic for the MRDviz simulation functionality
}
