% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{summarize_events}
\alias{summarize_events}
\title{Summarize Events}
\usage{
summarize_events(sim_data)
}
\arguments{
\item{sim_data}{Simulated data list}
}
\value{
A summary object
}
\description{
Creates a summary of survival events
}
