% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{summarize_trajectories}
\alias{summarize_trajectories}
\title{Summarize Trajectories}
\usage{
summarize_trajectories(sim_data)
}
\arguments{
\item{sim_data}{Simulated data list}
}
\value{
A summary object
}
\description{
Creates a summary of MRD trajectories
}
