% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_corr_matrix}
\alias{validate_corr_matrix}
\title{Validate Correlation Matrix}
\usage{
validate_corr_matrix(x)
}
\arguments{
\item{x}{Matrix to validate}
}
\value{
Validated correlation matrix
}
\description{
Check x is a valid correlation matrix
}
\keyword{internal}
