% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_family}
\alias{validate_family}
\title{Validate Family}
\usage{
validate_family(f)
}
\arguments{
\item{f}{The family argument specified by user}
}
\value{
Validated family object
}
\description{
Check and validate family argument
}
\keyword{internal}
