% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{verify_distribution}
\alias{verify_distribution}
\title{Verify Distribution of Time Points}
\usage{
verify_distribution(tij, ids, eot_time)
}
\arguments{
\item{tij}{Time points}

\item{ids}{Subject IDs}

\item{eot_time}{End of treatment time}
}
\value{
TRUE if distribution is valid
}
\description{
Verifies that time points are properly distributed
}
\keyword{internal}
