% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_viz.R
\name{viz_server}
\alias{viz_server}
\title{MRDviz Visualization Server Logic}
\usage{
viz_server(input, output, session, shared_storage)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{shared_storage}{Reactive values object for sharing data between tabs}
}
\description{
Server logic for the MRD visualization tab
}
